requirements.txt
numpy==1.19.1
opencv_python==4.3.0.36
torchvision==0.5.0
torch==1.4.0
matplotlib==3.3.1
ruamel.base==1.0.0
scipy==1.5.2
seaborn==0.10.1
